// src/schemas/statBaseSchemas.ts
import { z } from "zod/v4";

export const statBaseSchema = z.object({
  str: z.number().int().optional(),
  dex: z.number().int().optional(),
  con: z.number().int().optional(),
  int: z.number().int().optional(),
  wis: z.number().int().optional(),
  cha: z.number().int().optional(),
  fortitude: z.number().int().optional(),
  reflex: z.number().int().optional(),
  will: z.number().int().optional(),
  speedBonus: z.number().int().optional(),
  speedMax: z.number().int().optional(),
  speedMod: z.number().int().optional(),
  size: z.number().int().optional(),
  acArmor: z.number().int().optional(),
  acShield: z.number().int().optional(),
  acNatural: z.number().int().optional(),
  acDeflection: z.number().int().optional(),
  acMisc: z.number().int().optional(),
  acDexterityAvailable: z.boolean().optional(),
  acDexterityMax: z.number().int().optional(),
  skillPenalty: z.number().int().optional(),
  attackMelee: z.number().int().optional(),
  attackRanged: z.number().int().optional(),
  attackTouch: z.number().int().optional(),
  attackGrapple: z.number().int().optional(),
  damageMelee: z.number().int().optional(),
  damageRanged: z.number().int().optional(),
  hpTemp: z.number().int().optional(),
  damageReduction: z.number().int().optional(),
  damageReductionBypasses: z.string().optional(),
  resistCritChance: z.number().int().optional(),
  energyResistAcid: z.number().int().optional(),
  energyResistCold: z.number().int().optional(),
  energyResistElectricity: z.number().int().optional(),
  energyResistFire: z.number().int().optional(),
  energyResistSonic: z.number().int().optional(),
  energyResistForce: z.number().int().optional(),
  energyResistNegative: z.number().int().optional(),
  energyResistPositive: z.number().int().optional(),
  poisonResistance: z.number().int().optional(),
  charmResistance: z.number().int().optional(),
  fearResistance: z.number().int().optional(),
  sleepResistance: z.number().int().optional(),
  diseaseResistance: z.number().int().optional(),
  mindAffectingResistance: z.number().int().optional(),
  deathResistance: z.number().int().optional(),
  paralysisResistance: z.number().int().optional(),
  fatigueResistance: z.number().int().optional(),
  spellResistance: z.number().int().optional(),
  psionicResistance: z.number().int().optional(),
  casterLevelModifier: z.number().int().optional(),
  skillsBonuses: z.array(
    z.object({
      id: z.number().int().optional(),
      skillId: z.number().int(),
      bonus: z.number().int(),
    })
  ),
});